+++ blockstm/txnrunner.go
@@ -72,7 +72,7 @@ func (e STMRunner) Run(ctx context.Context, ms storetypes.MultiStore, txs [][]by
 		memTxs, estimates = preEstimates(txs, e.workers, authStore, bankStore, e.coinDenom(ms), e.txDecoder)
 	}
 
-	if err := ExecuteBlockWithEstimates(
+	err := ExecuteBlockWithEstimates(
 		ctx,
 		blockSize,
 		index,
@@ -99,7 +99,18 @@ func (e STMRunner) Run(ctx context.Context, ms storetypes.MultiStore, txs [][]by
 				incarnationCache[txn].Store(v)
 			}
 		},
-	); err != nil {
+	)
+
+	// fallback to sequential execution if parallel execution failed with scheduler error
+	if err != nil && err.Error() == "scheduler did not complete" {
+		results = make([]*abci.ExecTxResult, blockSize)
+		for i := range txs {
+			results[i] = deliverTx(txs[i], ms, i, nil)
+		}
+		return results, nil
+	}
+
+	if err != nil {
 		return nil, err
 	}